Move flatten component
