+++ tests/system/general/test_01_configuration_files.py
@@ -79,12 +79,6 @@ def q2dtest(add_app):
 
 
 @pytest.fixture(scope="class")
-def icepak_b(add_app):
-    app = add_app(project_name=ipk_name + "_b", application=Icepak)
-    return app
-
-
-@pytest.fixture(scope="class")
 def hfss3dl_a(add_app):
     app = add_app(project_name=diff_proj_name, application=Hfss3dLayout, subfolder=test_subfolder)
     return app
@@ -181,65 +175,6 @@ class TestClass:
         assert q2dtest.configurations.options.import_object_properties
         assert q2dtest.configurations.options.import_parametrics
         app.close_project(save=False)
-
-    @pytest.mark.skipif(
-        config["desktopVersion"] < "2023.1" and config["use_grpc"],
-        reason="Not working in 2022.2 gRPC",
-    )
-    def test_04b_icepak(self, icepak_b, add_app):
-        box1 = icepak_b.modeler.create_box([0, 0, 0], [10, 10, 10])
-        box1.surface_material_name = "Shellac-Dull-surface"
-        region = icepak_b.modeler["Region"]
-        icepak_b.monitor.assign_point_monitor_in_object(box1.name)
-        icepak_b.monitor.assign_face_monitor(box1.faces[0].id)
-        icepak_b.monitor.assign_point_monitor([5, 5, 5])
-        icepak_b.assign_openings(air_faces=region.bottom_face_x.id)
-        icepak_b.create_setup()
-        icepak_b.modeler.create_coordinate_system([10, 1, 10])
-        icepak_b.mesh.assign_mesh_region([box1.name])
-        icepak_b.mesh.global_mesh_region.MaxElementSizeX = "2mm"
-        icepak_b.mesh.global_mesh_region.MaxElementSizeY = "3mm"
-        icepak_b.mesh.global_mesh_region.MaxElementSizeZ = "4mm"
-        icepak_b.mesh.global_mesh_region.MaxSizeRatio = 2
-        icepak_b.mesh.global_mesh_region.UserSpecifiedSettings = True
-        icepak_b.mesh.global_mesh_region.UniformMeshParametersType = "XYZ Max Sizes"
-        icepak_b.mesh.global_mesh_region.MaxLevels = 2
-        icepak_b.mesh.global_mesh_region.BufferLayers = 1
-        icepak_b.mesh.global_mesh_region.update()
-        cs = icepak_b.modeler.create_coordinate_system(name="useless")
-        cs.props["OriginX"] = 20
-        cs.props["OriginY"] = 20
-        cs.props["OriginZ"] = 20
-        icepak_b.create_dataset(
-            "test_dataset",
-            [1, 2, 3, 4],
-            [1, 2, 3, 4],
-            z=None,
-            v=None,
-            is_project_dataset=False,
-            x_unit="cel",
-            y_unit="W",
-            v_unit="",
-        )
-        filename = icepak_b.design_name
-        icepak_b.export_3d_model(filename, icepak_b.working_directory, ".x_b", [], [])
-        fan = icepak_b.create_fan("test_fan")
-        icepak_b.modeler.user_defined_components[fan.name].move([1, 2, 3])
-        fan2 = icepak_b.modeler.user_defined_components[fan.name].duplicate_along_line([4, 5, 6])
-        icepak_b.modeler.user_defined_components[fan.name].rotate("Y")
-        fan3 = icepak_b.modeler.user_defined_components[fan.name].duplicate_around_axis("Z")
-        icepak_b.modeler.user_defined_components[fan.name].move([1, 2, 3])
-        fan4 = icepak_b.modeler.user_defined_components[fan.name].duplicate_around_axis("Z")
-        icepak_b.modeler.user_defined_components[fan2[0]].duplicate_and_mirror([4, 5, 6], [1, 2, 3])
-        conf_file = icepak_b.configurations.export_config()
-        assert icepak_b.configurations.validate(conf_file)
-        file_path = os.path.join(icepak_b.working_directory, filename + ".x_b")
-        app = add_app(application=Icepak, project_name="new_proj_Ipk", just_open=True)
-        app.modeler.import_3d_cad(file_path)
-        out = app.configurations.import_config(conf_file)
-        assert isinstance(out, dict)
-        assert app.configurations.validate(out)
-        assert app.configurations.results.global_import_success
 
     def test_05a_hfss3dlayout_setup(self, hfss3dl_a, local_scratch):
         setup2 = hfss3dl_a.create_setup("My_HFSS_Setup_2")  # Insert a setup.
